third_party/WebKit/LayoutTests/inspector-protocol/css/css-set-rule-selector.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="css-protocol-test.js"></script>
<script type="text/javascript">
function test()
{
    InspectorTest.eventHandler["CSS.styleSheetAdded"] = styleSheetAdded;
    InspectorTest.sendCommandOrDie("DOM.enable", {});
    InspectorTest.sendCommandOrDie("CSS.enable", {});

    var setRuleSelector;
    var verifyProtocolError;
    var dumpStyleSheet;

    var firstSelectorRange = {
        startLine: 0,
        startColumn: 0,
        endLine: 0,
        endColumn: 1
    };
    var secondSelectorRange = {
        startLine: 4,
        startColumn: 0,
        endLine: 4,
        endColumn: 18
    };
    var thirdSelectorRange = {
        startLine: 12,
        startColumn: 0,
        endLine: 12,
        endColumn: 14
    };
    var forthSelectorRange = {
        startLine: 12,
        startColumn: 36,
        endLine: 13,
        endColumn: 4
    };

    function styleSheetAdded(result)
    {
        var styleSheetId = result.params.header.styleSheetId;
        setRuleSelector = InspectorTest.setRuleSelector.bind(InspectorTest, styleSheetId, false);
        verifyProtocolError = InspectorTest.setRuleSelector.bind(InspectorTest, styleSheetId, true);
        dumpStyleSheet = InspectorTest.dumpStyleSheetText.bind(null, styleSheetId);
        InspectorTest.sendCommandOrDie("CSS.getStyleSheetText", { styleSheetId: styleSheetId }, onInitialStyleSheetText);
    }

    function onInitialStyleSheetText(result)
    {
        InspectorTest.log("==== Initial style sheet text ====");
        InspectorTest.log(result.text);
        InspectorTest.runTestSuite(testSuite);
    }

    var testSuite = [
        function testEditSimpleSelector(next)
        {
            setRuleSelector({
                range: firstSelectorRange,
                selector: ".EDITED-SELECTOR"
            }, InspectorTest.undoAndNext(next));
        },

        function testEditCopmlexSelector(next)
        {
            setRuleSelector({
                range: secondSelectorRange,
                selector: ".EDITED-SELECTOR:first-line"
            }, InspectorTest.undoAndNext(next));
        },

        function testEditSequentially(next)
        {
            var newSelectorText = ".EDITED";
            var oldSelectorLength = thirdSelectorRange.endColumn - thirdSelectorRange.startColumn;
            var lengthDelta = newSelectorText.length - oldSelectorLength;
            setRuleSelector({
                range: thirdSelectorRange,
                selector: newSelectorText,
            }, step2);

            function step2()
            {
                var range = {
                    startLine: forthSelectorRange.startLine,
                    startColumn: forthSelectorRange.startColumn + lengthDelta,
                    endLine: forthSelectorRange.endLine,
                    endColumn: forthSelectorRange.endColumn
                };
                setRuleSelector({
                    range: range,
                    selector: ".EDITED-2"
                }, InspectorTest.undoAndNext(next));
            }
        },

        function testEditSubSelector(next)
        {
            verifyProtocolError({
                range: { startLine: 12, startColumn: 8, endLine: 12, endColumn: 14 },
                selector: "line"
            }, next);
        },

        function testInvalidParameters(next)
        {
            verifyProtocolError({
                range: { startLine: "three", startColumn: 0, endLine: 4, endColumn: 0 },
                selector: "no matter what is here"
            }, next);
        },

        function testNegativeRangeParameters(next)
        {
            verifyProtocolError({
                range: { startLine: -1, startColumn: -1, endLine: 1, endColumn: 0 },
                selector: "html > body > div"
            }, next);
        },

        function testStartLineOutOfBounds(next)
        {
            verifyProtocolError({
                range: { startLine: 100, startColumn: 0, endLine: 100, endColumn: 0 },
                selector: "html > body > div"
            }, next);
        },

        function testEndLineOutOfBounds(next)
        {
            verifyProtocolError({
                range: { startLine: 0, startColumn: 0, endLine: 100, endColumn: 0 },
                selector: "html > body > div"
            }, next);
        },

        function testStartColumnBeyondLastLineCharacter(next)
        {
            verifyProtocolError({
                range: { startLine: 3, startColumn: 1000, endLine: 3, endColumn: 1000 },
                selector: "html > body > div"
            }, next);
        },

        function testEndColumnBeyondLastLineCharacter(next)
        {
            verifyProtocolError({
                range: { startLine: 3, startColumn: 0, endLine: 3, endColumn: 1000 },
                selector: "html > body > div"
            }, next);
        },

        function testInsertBeyondLastCharacterOfLastLine(next)
        {
            verifyProtocolError({
                range: { startLine: 3, startColumn: 2, endLine: 3, endColumn: 2 },
                selector: "html > body > div"
            }, next);
        },

        function testReversedRange(next)
        {
            verifyProtocolError({
                range: { startLine: 2, startColumn: 0, endLine: 0, endColumn: 0 },
                selector: "html > body > div"
            }, next);
        },
   ];
}

</script>
<link rel="stylesheet" href="resources/set-rule-selector.css"/>
</head>
<body onload="runTest();">
</body>
</html>
