third_party/WebKit/LayoutTests/inspector-protocol/css/css-set-rule-selector-expected.txt

==== Initial style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter { font-size: 200%; } div,
body {
    position: relative;
}



Running test: testEditSimpleSelector
==== Style sheet text ====
.EDITED-SELECTOR {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter { font-size: 200%; } div,
body {
    position: relative;
}



Running test: testEditCopmlexSelector
==== Style sheet text ====
* {
    box-sizing: border-box;
}

.EDITED-SELECTOR:first-line {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter { font-size: 200%; } div,
body {
    position: relative;
}



Running test: testEditSequentially
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

.EDITED { font-size: 200%; } div,
body {
    position: relative;
}


==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

.EDITED { font-size: 200%; } .EDITED-2 {
    position: relative;
}



Running test: testEditSubSelector
Expected protocol error: Source range didn't match any rule selector source range

Running test: testInvalidParameters
Expected protocol error: range.startLine must be a non-negative integer

Running test: testNegativeRangeParameters
Expected protocol error: range.startLine must be a non-negative integer

Running test: testStartLineOutOfBounds
Expected protocol error: Specified range is out of bounds

Running test: testEndLineOutOfBounds
Expected protocol error: Specified range is out of bounds

Running test: testStartColumnBeyondLastLineCharacter
Expected protocol error: Specified range is out of bounds

Running test: testEndColumnBeyondLastLineCharacter
Expected protocol error: Specified range is out of bounds

Running test: testInsertBeyondLastCharacterOfLastLine
Expected protocol error: Specified range is out of bounds

Running test: testReversedRange
Expected protocol error: Range start must not succeed its end
